+++ components/sponsored_benefits/app/controllers/sponsored_benefits/organizations/plan_design_proposals/carriers_controller.rb
@@ -1,7 +1,8 @@
 module SponsoredBenefits
   module Organizations
     class PlanDesignProposals::CarriersController < ApplicationController
-      include SponsoredBenefits::Organizations::PlanDesignHelpers
+      include SponsoredBenefits::Organizations::PlanDesigners
+      include SponsoredBenefits::Organizations::BenefitPresentationHelpers
 
       def index
         @carrier_names = ::Organization.load_carriers(
