+++ components/sponsored_benefits/app/controllers/sponsored_benefits/organizations/plan_design_proposals/contributions_controller.rb
@@ -1,7 +1,8 @@
 module SponsoredBenefits
   module Organizations
     class PlanDesignProposals::ContributionsController < ApplicationController
-      include SponsoredBenefits::Organizations::PlanDesignHelpers
+      include SponsoredBenefits::Organizations::PlanDesigners
+      include SponsoredBenefits::Organizations::BenefitGroupAccessors
 
       def index
         @plan = ::Plan.find(benefit_group.reference_plan_id)
