+++ components/sponsored_benefits/app/controllers/sponsored_benefits/organizations/plan_design_proposals/plan_reviews_controller.rb
@@ -1,7 +1,9 @@
 module SponsoredBenefits
   module Organizations
     class PlanDesignProposals::PlanReviewsController < ApplicationController
-      include SponsoredBenefits::Organizations::PlanDesignHelpers
+      include SponsoredBenefits::Organizations::PlanDesigners
+      include SponsoredBenefits::Organizations::BenefitGroupAccessors
+      include SponsoredBenefits::Organizations::BenefitPresentationHelpers
 
       def new
         @plan_design_organization = plan_design_organization
