+++ components/sponsored_benefits/app/controllers/sponsored_benefits/organizations/plan_design_proposals/plan_selections_controller.rb
@@ -1,7 +1,7 @@
 module SponsoredBenefits
   module Organizations
     class PlanDesignProposals::PlanSelectionsController < ApplicationController
-      include SponsoredBenefits::Organizations::PlanDesignHelpers
+      include SponsoredBenefits::Organizations::PlanDesigners
 
       before_action :published_plans_are_view_only
